feature - jar file acts as a self-contained app
Your app is already using the H2 in-memory database. This change loads
your test data from `book-data.sql` when the app starts.

It is my thought that this is a better configuration for showing off
your app. The `jar` file is a self-contained app which already contains
data to play with. This does not require the user to configure a pre-req
database to make the application work.

The downside is that the data is not persistent. Once the app is
stopped, the new data is lost. Restarting the app will reload the
test data again.